+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Environment protection Changelog
 
+## 1.1.0 - 2019-08-11
+### Added
+- Added cookie to allow visitors that enter a password or were in the whitelist. ([#2](https://github.com/JorgeAnzola/environment-protection/issues/2)) 
+
+### Fixed
+- Exit app when there is no redirect URL or template.
+
 ## 1.0.2 - 2019-08-08
 ### Fixed
 - Corrected validation for console requests.